Move LinkedGraphNode to its own file
