+++ fedot/core/optimisers/graph.py
@@ -1,7 +1,7 @@
 from typing import List, Union
 
 from fedot.core.dag.graph_delegate import GraphDelegate
-from fedot.core.dag.graph_node import LinkedGraphNode
+from fedot.core.dag.linked_graph_node import LinkedGraphNode
 from fedot.core.log import default_log
 
 OptNode = LinkedGraphNode
