+++ test/unit/pipelines/test_node.py
@@ -4,7 +4,7 @@ from sklearn.datasets import load_iris
 from sklearn.linear_model import LogisticRegression
 from sklearn.metrics import accuracy_score, classification_report
 
-from fedot.core.dag.graph_node import LinkedGraphNode
+from fedot.core.dag.linked_graph_node import LinkedGraphNode
 from fedot.core.dag.graph_utils import ordered_subnodes_hierarchy
 from fedot.core.data.data import InputData
 from fedot.core.data.data_split import train_test_data_setup
